HQLAA initial beta
Hybrid high-Quality Luma-adaptive Anti-Aliasing is a continuation of HQAA with customized FXAA routines built to select the strongest color channel to use as luma in order to catch fringe case aliasing in areas of a scene that are pure blue, pure red, or a mix of either. Typically, this is rare in most games, but some especially bright or saturated games can produce such scenes.