v2.10 update
FXAA blur is now sufficiently under control following the changes in the v2.9 series to allow 1080p to run an FXAA coarse pass. I had previously disabled it because the output was too blurry to realistically use. If you find 2.10 too blurry compared to v2.9, try using a lower preset or (if on custom mode) turn down the subpixel effect strength.