v14.0 update
So I discovered ReShade has a built-in IDE. This version is a full audit done using the shader's debug modes to fine tune parts that are working well and weed out parts that are not. This version is slimmer, faster (to compile at least), and a LOT more generally accurate.